separate file for matrix_iterator
